refactor: batch에서 job별로 공통으로 쓰이는 NxNy 클래스를 common 하위로 이동
