+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/repository/DashboardTodayWeatherRepository.java
@@ -2,7 +2,6 @@ package com.ImSnacks.NyeoreumnagiBatch.common.repository;
 
 import com.ImSnacks.NyeoreumnagiBatch.common.entity.DashboardTodayWeather;
 import com.ImSnacks.NyeoreumnagiBatch.common.entity.DashboardTodayWeatherId;
-import com.ImSnacks.NyeoreumnagiBatch.ultraviolet.writer.NxNy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
